Introduce the ResourceType concept
The "ResourceType" object is meant to replace the "type" field in the resource.
The reason to name it "ResourceType" is that "type" is a reserved word in python
and will lead to annoying situations and variable names down the road on every
level.

The ResourceType object is meant to hold information and capabilities regarding
a given type of resource.  Upcoming attributes will be a list of exporters/synchronizers
and data fields definitions

The ResourceTypeFactory is expected to be initialized at app launch by registering all
the resource types this instance will support, read from a config source.  The protection
against 2 resource types of the same name is to fail early in case a type would appear
twice in a configuration.

This factory could also be populated at run time via API but this is another topic
for another time.